requirements.txt
torch
torchvision
onnxruntime
Pillow
